+++ man-roxygen/funList.R
@@ -1,3 +1,4 @@
 #' @param funList list of character strings specifying functions to use for calculating
-#'                patch statistics (e.g., `list("landscapemetrics::lsm_l_area_mn")`).
-#'                These functions must work analogously to those in \pkg{landscapemetrics}.
+#' landscape or patch statistics (e.g., `list("lsm_l_area_mn")`).
+#' These functions must work analogously to those in \pkg{landscapemetrics}.
+#' See [default_landscape_metrics()], [default_patch_metrics()], [default_patch_metrics_seral()].